Merged PR 480: AppConfigHelper has been revamped
# Types of changes
<!-- Please put an `x` (e.g. [x]) in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)

# Checklist:
<!-- Please put an `x` (e.g. [x]) in all the boxes that apply: -->
- [x] All new and existing tests passed.
- [x] The code follows the code style and conventions of this project.
- [ ] The change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.

# Description of the change
<!-- Please provide enough information so others can review your pull request -->
Updating the AppConfigHelper to now create a new configuration from all current values in app config. This will occur any time a call is made to get a value from app config.

# Motivation for the change
<!-- Please explain the motivation for making this change -->
Fixes the issue where any values added to app config after the storage adapter is started do not appear in configuration (and fixes a similar affect for when app config values are deleted after starting the service).