fix: CA2234: Pass System.Uri objects instead of strings
In addition, a small tweak to the exception throwing. Throwing and catching the general-purpose Exception type is generally considered bad practice, and in this case, it was unnecessary because we could simply check for an empty/null webHookUrl and throw the exception we want straight away, outside of the try/catch.